@@ -19,7 +19,7 @@ authors = [
     {name = "Michele Ceriotti"}
 ]
 readme = "README.rst"
-requires-python = ">=3.8"
+requires-python = ">=3.10"
 license = {text = "BSD-3-Clause"}
 classifiers = [
     "Development Status :: 4 - Beta",
@@ -31,20 +31,20 @@ classifiers = [
     "Operating System :: MacOS :: MacOS X",
     "Operating System :: Microsoft :: Windows",
     "Programming Language :: Python :: 3",
-    "Programming Language :: Python :: 3.8",
     "Programming Language :: Python :: 3.9",
     "Programming Language :: Python :: 3.10",
     "Programming Language :: Python :: 3.11",
+    "Programming Language :: Python :: 3.12",
     "Topic :: Scientific/Engineering",
 ]
 dependencies = [
-    "scikit-learn>=1.1.0",
+    "scikit-learn >= 1.6.0",
+    "scipy >= 1.15.0",  # explicit to adhere to scikit-learn dependencies
 ]
 dynamic = ["version"]
 
 [project.optional-dependencies]
 examples = [
-    "ase",
     "matplotlib",
     "pandas",
     "tqdm",
@@ -75,9 +75,6 @@ include = [
 [tool.coverage.xml]
 output = 'tests/coverage.xml'
 
-[tool.pytest.ini_options]
-testpaths = "tests"
-
 [tool.isort]
 skip = "__init__.py"
 profile = "black"
@@ -86,3 +83,41 @@ indent = 4
 include_trailing_comma = true
 lines_after_imports = 2
 known_first_party = "skmatter"
+
+[tool.pytest.ini_options]
+testpaths = ["tests"]
+addopts = [
+    "--cov",
+    "--cov-append",
+    "--cov-report=",
+    "--import-mode=append",
+]
+
+[tool.ruff]
+exclude = ["docs/src/examples/"]
+lint.ignore = [
+    "F401",
+    "E203",
+    "D100",
+    "D101",
+    "D102",
+    "D205",
+    "D400",
+    "D401",
+]
+line-length = 88
+lint.select = [
+    "D",
+    "E",
+    "F",
+    "W",
+]
+
+[tool.ruff.lint.pydocstyle]
+convention = "numpy"
+
+[tool.ruff.lint.per-file-ignores]
+"examples/**" = [
+    "D205",
+    "D400",
+]
