@@ -1,11 +1,20 @@
 from os.path import dirname, join
 
 import numpy as np
-from sklearn.utils import Bunch, check_pandas_support
+import sklearn
+
+
+if sklearn.__version__ >= "1.5.0":
+    from sklearn.utils._optional_dependencies import check_pandas_support
+else:
+    from sklearn.utils import check_pandas_support
+
+from sklearn.utils import Bunch
 
 
 def load_nice_dataset():
     """Load and returns NICE dataset.
+
     Returns
     -------
     nice_data : sklearn.utils.Bunch
@@ -16,7 +25,6 @@ def load_nice_dataset():
       DESCR: `str` --
         The full description of the dataset.
     """
-
     module_path = dirname(__file__)
     target_filename = join(module_path, "data", "nice_dataset.npz")
     raw_data = np.load(target_filename)
@@ -92,6 +100,7 @@ def load_csd_1000r(return_X_y=False):
 
 def load_who_dataset():
     """Load and returns WHO dataset.
+
     Returns
     -------
     who_dataset : sklearn.utils.Bunch
@@ -100,7 +109,6 @@ def load_who_dataset():
                   as a Pandas dataframe.
           DESCR: `str` -- The full description of the dataset.
     """
-
     module_path = dirname(__file__)
     target_filename = join(module_path, "data", "who_dataset.csv")
     pd = check_pandas_support("load_who_dataset")
@@ -111,33 +119,54 @@ def load_who_dataset():
 
 
 def load_roy_dataset():
-    """Load and returns the ROY dataset, which contains structures,
-    energies and SOAP-derived descriptors for 264 polymorphs of ROY,
+    """Load and returns the ROY dataset, which contains densities,
+    energies and SOAP-derived descriptors for 264 structures of polymorphs of ROY,
     from [Beran et Al, Chemical Science (2022)](https://doi.org/10.1039/D1SC06074K)
+    Each structure is labeled as "Known" or "Unknown".
 
     Returns
     -------
     roy_dataset : sklearn.utils.Bunch
       Dictionary-like object, with the following attributes:
-          structures : `ase.Atoms` -- the roy structures as ASE objects
-          features: `np.array` -- SOAP-derived descriptors for the structures
-          energies: `np.array` -- energies of the structures
+          densities : `np.array` -- the densities of the structures
+          structure_types : `np.array` -- the type of the structures
+          features : `np.array` -- SOAP-derived descriptors for the structures
+          energies : `np.array` -- energies of the structures
     """
-
     module_path = dirname(__file__)
-    target_structures = join(module_path, "data", "beran_roy_structures.xyz.bz2")
+    target_properties = join(module_path, "data", "beran_roy_properties.npz")
+    properties = np.load(target_properties)
 
-    try:
-        from ase.io import read
-    except ImportError:
-        raise ImportError("load_roy_dataset requires the ASE package.")
+    return Bunch(
+        densities=properties["densities"],
+        energies=properties["energies"],
+        structure_types=properties["structure_types"],
+        features=properties["feats"],
+    )
 
-    import bz2
 
-    structures = read(bz2.open(target_structures, "rt"), ":", format="extxyz")
-    energies = np.array([f.info["energy"] for f in structures])
+def load_hbond_dataset():
+    """Load and returns the hydrogen bond dataset, which contains
+    a set of 3D descriptors for 27233 hydrogen bonds and corresponding
+    weights, from [Gasparotto et Al, The Journal of Chemical Physics]
+    (https://doi.org/10.1063/1.4900655)
 
-    target_features = join(module_path, "data", "beran_roy_features.npz")
-    features = np.load(target_features)["feats"]
+    Returns
+    -------
+    hbond_dataset : sklearn.utils.Bunch
+      Dictionary-like object, with the following attributes:
+          descriptors : `numpy.ndarray` -- the descriptors of hydrogen bond dataset
+          weights : `numpy.ndarray` -- the weights of each sample in the dataset
+    """
+    module_path = dirname(__file__)
+    target_filename = join(module_path, "data", "h2o-blyp-piglet.npz")
+    raw_data = np.load(target_filename)
 
-    return Bunch(structures=structures, features=features, energies=energies)
+    with open(join(module_path, "descr", "h2o-blyp-piglet.rst")) as rst_file:
+        fdescr = rst_file.read()
+
+    return Bunch(
+        descriptors=raw_data["descriptors"],
+        weights=raw_data["weights"],
+        DESCR=fdescr,
+    )
