@@ -1,6 +1,6 @@
 from sklearn.utils.estimator_checks import parametrize_with_checks
 
-from skmatter.decomposition import KernelPCovR, PCovR
+from skmatter.decomposition import KernelPCovR, PCovC, PCovR
 from skmatter.feature_selection import CUR as fCUR
 from skmatter.feature_selection import FPS as fFPS
 from skmatter.feature_selection import PCovCUR as fPCovCUR
@@ -13,6 +13,7 @@ from skmatter.preprocessing import KernelNormalizer, StandardFlexibleScaler
     [
         KernelPCovR(mixing=0.5),
         PCovR(mixing=0.5),
+        PCovC(mixing=0.5),
         fCUR(),
         fFPS(),
         fPCovCUR(),
@@ -23,4 +24,5 @@ from skmatter.preprocessing import KernelNormalizer, StandardFlexibleScaler
     ]
 )
 def test_sklearn_compatible_estimator(estimator, check):
+    """Test of the estimators are compatible with sklearn."""
     check(estimator)
