@@ -18,10 +18,7 @@ class TestCUR(unittest.TestCase):
             _ = selector.transform(self.X)
 
     def test_restart(self):
-        """
-        This test checks that the model can be restarted with a new instance
-        """
-
+        """Check that the model can be restarted with a new instance."""
         ref_selector = CUR(n_to_select=self.X.shape[-1] - 3).fit(X=self.X)
         ref_idx = ref_selector.selected_idx_
 
@@ -34,9 +31,7 @@ class TestCUR(unittest.TestCase):
             self.assertEqual(selector.selected_idx_[i], ref_idx[i])
 
     def test_non_it(self):
-        """
-        This test checks that the model can be run non-iteratively
-        """
+        """Check that the model can be run non-iteratively."""
         C = self.X.T @ self.X
         _, UC = np.linalg.eigh(C)
         ref_idx = np.argsort(-(UC[:, -1] ** 2.0))[:-1]
