@@ -5,10 +5,7 @@ from skmatter.utils import get_progress_bar
 
 class PBarTest(unittest.TestCase):
     def test_no_tqdm(self):
-        """
-        This test checks that the model cannot use a progress bar when tqdm
-        is not installed
-        """
+        """Check that the model cannot use a progress bar when tqdm is not installed."""
         import sys
 
         sys.modules["tqdm"] = None
@@ -17,9 +14,8 @@ class PBarTest(unittest.TestCase):
             _ = get_progress_bar()
         self.assertEqual(
             str(cm.exception),
-            "tqdm must be installed to use a progress bar."
-            "Either install tqdm or re-run with"
-            "progress_bar = False",
+            "tqdm must be installed to use a progress bar. Either install tqdm or "
+            "re-run with progress_bar = False",
         )
 
 
